use float32 for combat calculations

+++ characters/stats.go
@@ -4,28 +4,28 @@ import (
 	"fmt"
 )
 
-func (stats Stats) CriticalValue() int {
-	return int(float64(stats.Agility) * float64(1.45)) - stats.Level
+func (stats Stats) CriticalValue() float32 {
+	return float64(stats.Agility) * 1.45 - float32(stats.Level)
 }
 
-func (stats Stats) DodgeValue() int {
-	return (stats.Agility * 1.9) - stats.Level
+func (stats Stats) DodgeValue() float32 {
+	return (float32(stats.Agility) * 1.9) - float32(stats.Level)
 }
 
-func (stats Stats) BlockValue() int {
-	return (stats.Strength * 1.9 + stats.Block) - stats.Level
+func (stats Stats) BlockValue() float32 {
+	return float32(stats.Strength) * 1.9 + float32(stats.Block) - float32(stats.Level)
 }
 
-func (stats Stats) AccuracyRating() int {
-	return (stats.Agility + stats.Expertise)
+func (stats Stats) AccuracyRating() float32 {
+	return float32(stats.Agility) + float32(stats.Expertise)
 }
 
-func (stats Stats) MaxHealth() int {
-	return (stats.Vitality * 8) + (stats.Strength * 2)
+func (stats Stats) MaxHealth() float32 {
+	return (float32(stats.Vitality) * 8) + (float32(stats.Strength) * 2)
 }
 
-func (stats Stats) MaxFocus() int {
-	return stats.Intelligence * 10
+func (stats Stats) MaxFocus() float32 {
+	return float32(stats.Intelligence) * 10
 }
 
 func (stats Stats) DisplayHealth() string {
